# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs DTU_DZ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/bai-viet/data": {
            "get": {
                "tags": [
                    "Bài Viết"
                ],
                "summary": "Bai Viet Data",
                "description": "Get Data Bai Viet",
                "operationId": "f24a31a98f8d4ea7e0c5c4b48cc0ebcd",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/api/admin/bai-viet/create": {
            "post": {
                "tags": [
                    "Bài Viết"
                ],
                "summary": "Create Bai Viet",
                "description": "",
                "operationId": "e48eb2891403f774f2bb5ac3d607544b",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "ten_bai_viet",
                                    "slug_bai_viet",
                                    "mo_ta",
                                    "tinh_trang",
                                    "hinh_anh"
                                ],
                                "properties": {
                                    "ten_bai_viet": {
                                        "type": "text"
                                    },
                                    "slug_bai_viet": {
                                        "type": "text"
                                    },
                                    "mo_ta": {
                                        "type": "text"
                                    },
                                    "tinh_trang": {
                                        "type": "number"
                                    },
                                    "hinh_anh": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/admin/bai-viet/status": {
            "post": {
                "tags": [
                    "Bài Viết"
                ],
                "summary": "Doi Trang Thai Bai Viet",
                "description": "",
                "operationId": "c63eca6d95fa5778441d737a2a5965f5",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/admin/bai-viet/delete": {
            "post": {
                "tags": [
                    "Bài Viết"
                ],
                "summary": "Xoa Bai Viet",
                "description": "",
                "operationId": "3f94b579f4d1781233ce042e45f93dd6",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/admin/bai-viet/update": {
            "post": {
                "tags": [
                    "Bài Viết"
                ],
                "summary": "Update Bai Viet",
                "description": "",
                "operationId": "2dcbe291a97bc0bdcfeee3db45271c73",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "ten_bai_viet",
                                    "slug_bai_viet",
                                    "mo_ta",
                                    "tinh_trang",
                                    "hinh_anh"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "ten_bai_viet": {
                                        "type": "text"
                                    },
                                    "slug_bai_viet": {
                                        "type": "text"
                                    },
                                    "mo_ta": {
                                        "type": "text"
                                    },
                                    "tinh_trang": {
                                        "type": "number"
                                    },
                                    "hinh_anh": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}